Outsource cross compilation process
The cross compilation process now lives in two separate operations.

Additionally I changed the way how to call a deployment operation. A
method `#configure` was added which yields the operation before
executing it. This makes it possible to apply properties that are not
defined in the main operation's constructor (for example properties of
nested operations).

Moreover, command stubs for the tests were added. This was needed to
mimic the bahaviour normal command execution would produce. For example
when compiling we expect the link command to be printed to stdout.

`STDOUT` was replaced with `DeployCR.stdout` which makes it possible to
change the value if needed (for example in specs).